Moved YarnClusterManager to yarn project.
