--- core/src/main/resources/META-INF/services/org.apache.spark.scheduler.ExternalClusterManager
@@ -1,1 +1,0 @@
-org.apache.spark.scheduler.YarnClusterManager
