+++ core/src/test/resources/META-INF/services/org.apache.spark.scheduler.ExternalClusterManager
@@ -1,2 +1,1 @@
 org.apache.spark.scheduler.DummyExternalClusterManager
-org.apache.spark.scheduler.YarnClusterManager
