+++ yarn/src/main/resources/META-INF/services/org.apache.spark.scheduler.ExternalClusterManager
@@ -1,0 +1,1 @@
+org.apache.spark.scheduler.cluster.YarnClusterManager
